+++ src/ares/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "0.5.0"
+__version__ = "0.9.0"
 
 from .main import *  # noqa: F403 F401