+++ src/ares/managers/path_manager.py
@@ -4,7 +4,14 @@
 from typing import Any, Dict, List, Union
 
 import numpy as np
-from consts import (
+from sc2.ids.effect_id import EffectId
+from sc2.ids.unit_typeid import UnitTypeId as UnitID
+from sc2.position import Point2
+from sc2.unit import Unit
+from scipy import spatial
+from scipy.spatial import KDTree
+
+from ares.consts import (
     ACTIVE_GRID,
     AIR,
     AIR_AVOIDANCE,
@@ -41,17 +48,10 @@ from consts import (
     ManagerName,
     ManagerRequestType,
 )
-from custom_bot_ai import CustomBotAI
-from dicts.weight_costs import WEIGHT_COSTS
-from managers.manager import Manager
-from managers.manager_mediator import IManagerMediator, ManagerMediator
-from sc2.ids.effect_id import EffectId
-from sc2.ids.unit_typeid import UnitTypeId as UnitID
-from sc2.position import Point2
-from sc2.unit import Unit
-from scipy import spatial
-from scipy.spatial import KDTree
-
+from ares.custom_bot_ai import CustomBotAI
+from ares.dicts.weight_costs import WEIGHT_COSTS
+from ares.managers.manager import Manager
+from ares.managers.manager_mediator import IManagerMediator, ManagerMediator
 from MapAnalyzer import MapData